Rename ucar.ma -> ucar.array. Implement others of sean's suggestions.
