+++ cdm/core/src/main/java/ucar/array/Index.java
@@ -2,13 +2,15 @@
  * Copyright (c) 1998-2018 University Corporation for Atmospheric Research/Unidata
  * See LICENSE for license information.
  */
-package ucar.ma;
+package ucar.array;
 
 import com.google.common.base.Preconditions;
 
 /**
  * Indexes for Multidimensional arrays. An Index refers to a particular element of an array.
  * This is a generalization of index as int[].
+ * TODO: Some of these methods, maybe the whole class is to ease the transition from earlier versions.
+ * TODO: evaluate how useful this is.
  */
 public class Index {
   private final int[] current; // current element's index
@@ -44,9 +46,10 @@ public class Index {
    * @param dim set this dimension
    * @param value to this value
    */
-  public void setDim(int dim, int value) {
+  public Index setDim(int dim, int value) {
     Preconditions.checkArgument(dim < current.length);
     current[dim] = value;
+    return this;
   }
 
   /** set current element at dimension 0 to v */
