+++ cdm/core/src/main/java/ucar/array/package.html
@@ -1,6 +1,6 @@
 <HTML>
 <HEAD>
-<TITLE>package ucar.ma</TITLE>
+<TITLE>package ucar.array</TITLE>
 </HEAD>
 <BODY>
 Experimental version of multidimensional arrays stored in memory.
