+++ cdm/core/src/test/java/ucar/array/TestStorage.java
@@ -2,7 +2,7 @@
  * Copyright (c) 1998-2020 John Caron and University Corporation for Atmospheric Research/Unidata
  * See LICENSE for license information.
  */
-package ucar.ma;
+package ucar.array;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.fail;
